bitmap class, window changes, create gameclient.hpp and gameserver.hpp, warning fixes, slight cleaning

+++ SBBB_Application/include/ChunkManager.hpp
@@ -35,7 +35,7 @@ public:
 	void regenVBOs();
 	void enqueueGen(ChunkPos p_chunkPos);
 	//bool genFromQueue();
-	void genFixed(size_t x, size_t y);
+	void genFixed(uint32_t x, uint32_t y);
 	bool autoGen(Camera& p_cam);
 
 	void startThreads();
